my_pipeline/requirement.txt
google-api-python-client==2.61.0
pathlib==1.0.1
google-cloud-storage==2.5.0
schedule==1.1.0
dnspython==2.2.1
pymongo==4.2.0
python-dotenv==0.21.0
Pillow==7.1.2
numpy==1.21.6
torchvision